Make BuildStarknetChainDriver context-generic
