+++ relayer/crates/starknet-integration-tests/src/lib.rs
@@ -1,6 +1,9 @@
 #![recursion_limit = "256"]
 
+extern crate alloc;
+
 pub mod contexts;
+pub mod impls;
 pub mod utils;
 
 #[cfg(test)]
